+++ raster/src/main/scala/geotrellis/raster/summary/polygonal/Implicits.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Azavea
+ * Copyright 2019 Azavea
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,16 +17,30 @@
 package geotrellis.raster.summary.polygonal
 
 import geotrellis.raster._
-
-
-object Implicits extends Implicits
+import geotrellis.util.GetComponent
 
 trait Implicits {
-  implicit class withSinglebandTilePolygonalSummaryMethods(
-    val self: Tile
-  ) extends SinglebandTilePolygonalSummaryMethods
+  implicit def rasterHasRasterExtent[T <: CellGrid[Int]]
+    : GetComponent[Raster[T], RasterExtent] =
+    new GetComponent[Raster[T], RasterExtent] {
+      override def get: Raster[T] => RasterExtent = _.rasterExtent
+    }
 
-  implicit class withMultibandTilePolygonalSummaryMethods(
-    val self: MultibandTile
-  ) extends MultibandTilePolygonalSummaryMethods
+  implicit class withPolygonalSummaryResultMethods[A](val self: PolygonalSummaryResult[A]) {
+    def toEither: Either[NoIntersection.type, A] = {
+      self match {
+        case NoIntersection => Left(NoIntersection)
+        case Summary(value) => Right(value)
+      }
+    }
+
+    def toOption: Option[A] = {
+      self match {
+        case NoIntersection => None
+        case Summary(value) => Some(value)
+      }
+    }
+  }
 }
+
+object Implicits extends Implicits