+++ docs/CHANGELOG.rst
@@ -5,6 +5,10 @@ Changelog
 -----
 
 API Changes & Project structure changes
+
+- ``geotrellis-raster``
+
+  - **Change:** ``geotrellis.raster.summary.polygonal.[Multi]TilePolygonalSummaryHandler`` replaced with ``geotrellis.raster.summary.polygonal.PolygonalSummary. Users should expect to implement concrete subclasses of ``geotrellis.raster.summary.CellVisitor`` and pass those to the new polygonalSummary methods. There are a number of default CellVisitor implementations provided for simple operations in `geotrellis.raster.summary.visitors``
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 - ``geotrellis.slick``